git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/branches/2.5@6096 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ gftool/gfprep/gfarm_dirtree.c
@@ -454,7 +454,7 @@ dents_loop:
 				fwrite(buf, 1, retv, to_parent);
 				if (ferror(to_parent))
 					goto dents_error;
-				if (retv < sizeof(buf))
+				if (retv < (int)sizeof(buf))
 					break; /* EOF */
 			}
 			fclose(tmpfp);
